refactor: isolate agent harness client state
