+++ src/lib/sidecar-ownership.ts
@@ -1,0 +1,6 @@
+export function harnessOwnershipCompatible(
+    agentHarnessEnabled: boolean,
+    nativeOwned: boolean,
+): boolean {
+    return !agentHarnessEnabled || nativeOwned;
+}
